Changed the lighting so that the directional light only has a slight effect and the ambient light is much brighter. Then updated all the models to include the same "burnt in" lighting as the original game.
Fixed a bug in the hit check for big explosion that could cause the game to run the hit check loop forever.

Made the launcher launch missiles from a more reasonable looking position.

Fixed an error in 12-bit to 24-bit colour conversion in smd_to_shp.py

Removed some obsolete files.

+++ src/big_explosions.h
@@ -22,7 +22,7 @@ public:
 class Boom: public MissileDeath
 {
 private:
-    float m_nextDamageCheckTime;
+    double m_nextDamageCheckTime;
     float GetSize();
 
 public:
